fix(requirements.txt): update and extend scope
By now, an update/move to `pyreadline3` is sensible (as already in
the principal branch).  Extension with `pytest` offers an optional
test on site.

In Linux Debian 13/trixie (virtual environment of Python 3.12.4 and
pytest 8.3.3), out of 540 tests all 540 pass, and 4 are tagged
with a warning.  Pylint (3.3.1)/astroid (3.3.5) indicate a couple of
possible improvements (r strings, etc.) and hence rate
test_appendfilename.py with 6.21/10 -- but does not report a syntax
error.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>