refactor: edit tests to smartly prepend a string
The third section of tests about smartly prepending a string -- i.e.
to insert a string right after the time or date stamp -- was edited.
Manual tests in Linux Debian and Windows 10 as well as with pytest
on either operating system confirm inconsistencies of the format
of the file name eventually generated.[1,2]

In addition, the implementation differs in robustness to deploy
special characters like for instance the underscore as a separator
to enclose the string to add.  Instead of `_` (as in Debian), Windows
tends to enclose these separators with single quotes (`'_'`) instead.
The default by `appendfilename` to use a single blank character
still is considered "safe to be used" _for `appendfilename`_ though
e.g. the underscore would represent a safe alternative if file names
with white space are not acceptable.

[1] https://github.com/novoid/appendfilename/issues/15
[2] https://github.com/novoid/appendfilename/issues/16

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>